[department-of-veterans-affairs/orchid#88] Refactors veteran information section and contact information page
